requirements.txt
groq>=0.20.0
python-dotenv>=1.0.1
sounddevice>=0.5.1
# Pinned <13 because we type-annotate against websockets.WebSocketClientProtocol,
# which was removed in the 13.x rewrite.
websockets>=12.0,<13.0
